backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai==1.3.0
python-docx==1.1.0
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.0 